refactor event handling by removing eventBus and integrating event processing directly into useEvents; improve event ID tracking to prevent duplicates
